pointer and autoscroll fixes
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@7545 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ gui/rosegardencanvasview.h
@@ -160,8 +160,8 @@ protected:
     int m_scrollDirectionConstraint;
     bool m_autoScrolling;
 
-    static const int DefaultSmoothScrollTimeInterval;
-    static const int DefaultMinDeltaScroll;
+    static const int    DefaultSmoothScrollTimeInterval;
+    static const double DefaultMinDeltaScroll;
 
     static const int AutoscrollMargin;
     static const int InitialScrollTime;
